megapp/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.21.2
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.1.0
pyparsing==3.1.4
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
